@@ -9,42 +9,52 @@
     "build:ssr": "ng run ScriptBeeWebApp:server:dev",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "run:all": "node node_modules/@angular-architects/module-federation/src/server/mf-dev-server.js"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.5",
-    "@angular/cdk": "^13.3.5",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/flex-layout": "^13.0.0-beta.36",
-    "@angular/forms": "~13.3.0",
-    "@angular/material": "^13.1.3",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
+    "@angular-architects/module-federation": "^14.3.10",
+    "@angular-devkit/core": "^14.2.2",
+    "@angular-devkit/schematics": "^14.2.2",
+    "@angular/animations": "14.2.1",
+    "@angular/cdk": "14.2.1",
+    "@angular/common": "14.2.1",
+    "@angular/compiler": "14.2.1",
+    "@angular/core": "14.2.1",
+    "@angular/flex-layout": "^14.0.0-beta.40",
+    "@angular/forms": "14.2.1",
+    "@angular/material": "14.2.1",
+    "@angular/platform-browser": "14.2.1",
+    "@angular/platform-browser-dynamic": "14.2.1",
+    "@angular/router": "14.2.1",
     "@materia-ui/ngx-monaco-editor": "^6.0.0",
-    "@mdi/angular-material": "^6.6.96",
+    "@mdi/angular-material": "^7.0.96",
+    "@ngrx/effects": "^14.3.1",
+    "@ngrx/entity": "^14.3.1",
+    "@ngrx/store": "^14.3.1",
+    "@schematics/angular": "^14.2.2",
     "angular-split": "^13.2.0",
-    "monaco-editor": "^0.33.0",
-    "rxjs": "~7.5.0",
+    "monaco-editor": "^0.34.0",
+    "ngx-signalr-websocket": "^0.4.1",
+    "rxjs": "~7.5.6",
     "tslib": "^2.4.0",
-    "zone.js": "^0.11.5"
+    "zone.js": "^0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.5",
-    "@angular-eslint/schematics": "^13.2.1",
-    "@angular/cli": "~13.3.5",
-    "@angular/compiler-cli": "~13.3.0",
-    "@types/jasmine": "^4.0.3",
-    "@types/node": "^17.0.31",
-    "jasmine-core": "^4.1.1",
-    "karma": "~6.3.0",
-    "karma-chrome-launcher": "~3.1.0",
+    "@angular-devkit/build-angular": "14.2.2",
+    "@angular-eslint/schematics": "14.0.4",
+    "@angular/cli": "14.2.2",
+    "@angular/compiler-cli": "14.2.1",
+    "@types/jasmine": "^4.3.0",
+    "@types/node": "18.7.16",
+    "jasmine-core": "^4.4.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.1.1",
     "karma-coverage": "^2.2.0",
-    "karma-jasmine": "^5.0.0",
-    "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.6.2"
+    "karma-jasmine": "^5.1.0",
+    "karma-jasmine-html-reporter": "2.0.0",
+    "ngx-build-plus": "^14.0.0",
+    "typescript": "~4.7.0"
   }
 }
