@@ -1,16 +1,16 @@
-import {Component, OnInit} from '@angular/core';
-import {FileSystemService} from '../../services/file-system/file-system.service';
-import {ProjectDetailsService} from '../project-details.service';
-import {FileTreeNode} from './fileTreeNode';
-import {ActivatedRoute, Router} from '@angular/router';
-import {CreateScriptDialogComponent} from './create-script-dialog/create-script-dialog.component';
-import {MatDialog} from '@angular/material/dialog';
-import {MatSnackBar} from '@angular/material/snack-bar';
-import {OutputFilesService} from '../../services/output/output-files.service';
-import {OutputFile} from '../../services/output/output-file';
-import {TreeNode} from '../../shared/tree-node';
-import {debounceTime, distinctUntilChanged, distinctUntilKeyChanged, filter, first} from 'rxjs/operators';
-import {RunScriptResult} from '../../services/run-script/run-script-result';
+import { Component, OnInit } from '@angular/core';
+import { FileSystemService } from '../../services/file-system/file-system.service';
+import { ActivatedRoute, Router } from '@angular/router';
+import { CreateScriptDialogComponent } from './create-script-dialog/create-script-dialog.component';
+import { MatDialog } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { OutputFile } from '../../services/output/output-file';
+import { filter } from 'rxjs/operators';
+import { Store } from "@ngrx/store";
+import { selectProjectDetails } from "../../state/project-details/project-details.selectors";
+import { fetchScriptTree } from "../../state/script-tree/script-tree.actions";
+import { selectScriptTreeLeafClick } from "../../state/script-tree/script-tree.selectors";
+import { selectLastRunOutput } from "../../state/outputs/output.selectors";
 
 @Component({
   selector: 'app-scripts-content',
@@ -18,89 +18,50 @@ import {RunScriptResult} from '../../services/run-script/run-script-result';
   styleUrls: ['./scripts-content.component.scss']
 })
 export class ScriptsContentComponent implements OnInit {
-
-  fileStructureTree: FileTreeNode[] = [];
-  consoleOutput = '';
-  outputErrors = '';
-  outputFiles: OutputFile[] = [];
   projectId = '';
-  runId = '';
 
-  constructor(private fileSystemService: FileSystemService, private projectDetailsService: ProjectDetailsService,
-              private router: Router, private route: ActivatedRoute, private dialog: MatDialog, private snackBar: MatSnackBar,
-              private outputFilesService: OutputFilesService) {
+  outputFiles: OutputFile[] = [];
+  runIndex: number | undefined;
+  outputErrorsId: string | undefined;
+  consoleOutputId: string | undefined;
+
+  constructor(private fileSystemService: FileSystemService, private router: Router,
+              private route: ActivatedRoute, private dialog: MatDialog,
+              private snackBar: MatSnackBar, private store: Store) {
   }
 
   ngOnInit(): void {
-    this.loadProjectFileStructure();
-    this.getOutput();
-  }
+    this.store.select(selectProjectDetails).pipe(filter(x => !!x)).subscribe({
+      next: projectDetails => {
+        this.projectId = projectDetails.data.projectId;
+        this.store.dispatch(fetchScriptTree({projectId: projectDetails.data.projectId}));
+      }
+    });
 
-  onLeafClick(node: TreeNode) {
-    const fileNode = node as FileTreeNode;
-    this.router.navigate([fileNode.srcPath], {relativeTo: this.route});
+    this.store.select(selectScriptTreeLeafClick).pipe(filter(x => !!x)).subscribe({
+      next: node => {
+        void this.router.navigate([node.id], {relativeTo: this.route});
+      }
+    });
+
+    this.store.select(selectLastRunOutput).subscribe({
+      next: outputState => {
+        this.outputErrorsId = outputState.results.filter(x => x.type == 'RunError').map(x => x.id)[0];
+        this.consoleOutputId = outputState.results.filter(x => x.type == 'Console').map(x => x.id)[0];
+        this.outputFiles = outputState.results.filter(x => x.type == 'File').map(x => ({
+          fileId: x.id,
+          fileName: x.name,
+        }));
+        this.runIndex = outputState.runIndex;
+      }
+    })
   }
 
   onCreateNewScriptButtonClick() {
-    const dialogRef = this.dialog.open(CreateScriptDialogComponent, {
+    this.dialog.open(CreateScriptDialogComponent, {
       width: '300px',
-      disableClose: true
-    });
-
-    dialogRef.afterClosed().subscribe(scriptPath => {
-      if (scriptPath) {
-        this.loadProjectFileStructure(() => {
-          const newScriptPath = scriptPath.replaceAll('/', '\\');
-          this.router.navigate([newScriptPath], {relativeTo: this.route});
-        });
-      }
-    }, (error: any) => {
-      this.snackBar.open('Could not create script!', 'Ok', {
-        duration: 4000
-      });
+      disableClose: true,
+      data: {projectId: this.projectId}
     });
   }
-
-  private getOutput() {
-    this.projectDetailsService.lastRunErrorMessage.subscribe(message => {
-      this.outputErrors = message;
-      this.consoleOutput = '';
-      this.outputFiles = [];
-console.log(message)
-      if (!message) {
-        this.projectDetailsService.lastRunResult.pipe(filter(x => x !== undefined)).subscribe(runResult => {
-          console.log(runResult);
-          if (runResult == null) {
-            return;
-          }
-          this.projectId = runResult.projectId;
-          this.runId = runResult.runId;
-          this.outputFilesService.getConsoleOutputContent(runResult.consoleOutputName).subscribe(consoleContent => {
-            this.consoleOutput = consoleContent;
-          });
-          this.outputFiles = runResult.outputFiles;
-        });
-      }
-    });
-  }
-
-  private loadProjectFileStructure(callback?: () => void) {
-    this.projectDetailsService.project.subscribe(project => {
-      if (project) {
-        this.fileSystemService.getFileSystem(project.projectId).subscribe(fileTreeNode => {
-          if (fileTreeNode) {
-            this.fileStructureTree = [fileTreeNode];
-            if (callback) {
-              callback();
-            }
-          }
-        });
-      }
-    });
-  }
-
-  private areRunResultsEqual(runRes1: RunScriptResult, runRes2: RunScriptResult) {
-    return runRes1 && runRes2 && runRes1.runId == runRes2.runId && runRes1.projectId == runRes2.projectId;
-  }
-
 }
