@@ -1,11 +1,12 @@
-import {Injectable} from '@angular/core';
-import {HttpClient} from '@angular/common/http';
-import {Project} from '../../projects/project';
-import {map} from 'rxjs/operators';
-import {Observable} from 'rxjs';
-import {contentHeaders} from '../../shared/headers';
-import {TreeNode} from '../../shared/tree-node';
-import {ReturnedProject} from './returned-project';
+import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+import { map } from 'rxjs/operators';
+import { Observable } from 'rxjs';
+import { contentHeaders } from '../../shared/headers';
+import { TreeNode } from '../../shared/tree-node';
+import { ReturnedNode, ReturnedProject } from './returned-project';
+import { ProjectData } from "../../state/project-details/project";
+import { ReturnedContextSlice } from "./returned-context-slice";
 
 @Injectable({
   providedIn: 'root'
@@ -17,18 +18,21 @@ export class ProjectService {
   constructor(private http: HttpClient) {
   }
 
-  getProject(projectId: string): Observable<Project> {
+  getProject(projectId: string): Observable<ProjectData> {
     return this.http.get<ReturnedProject>(`${this.projectsAPIUrl}/${projectId}`, {headers: contentHeaders})
       .pipe(map((data: ReturnedProject) => {
         return ProjectService.convertReturnedProjectToProject(data);
       }));
   }
 
-  getProjectContext(projectId: string) {
-    return this.http.get<TreeNode[]>(`${this.projectsAPIUrl}/context/${projectId}`, {headers: contentHeaders});
+  getProjectContext(projectId: string): Observable<TreeNode[]> {
+    return this.http.get<ReturnedContextSlice[]>(`${this.projectsAPIUrl}/context/${projectId}`, {headers: contentHeaders})
+      .pipe(map((data: ReturnedContextSlice[]) => {
+        return ProjectService.convertReturnedContextSlicesToContext(data);
+      }));
   }
 
-  getAllProjects(): Observable<Project[]> {
+  getAllProjects(): Observable<ProjectData[]> {
     return this.http.get<ReturnedProject[]>(this.projectsAPIUrl, {headers: contentHeaders}).pipe(map((data: ReturnedProject[]) => {
       return data.map((project: ReturnedProject) => {
         return ProjectService.convertReturnedProjectToProject(project);
@@ -47,24 +51,18 @@ export class ProjectService {
     return this.http.delete(`${this.projectsAPIUrl}/${projectId}`, {headers: contentHeaders});
   }
 
-  private static convertReturnedProjectToProject(returnedProject: ReturnedProject) {
-    const savedFiles: TreeNode[] = returnedProject.savedFiles.map(file => {
+  private static convertReturnedProjectToProject(returnedProject: ReturnedProject): ProjectData {
+    function mapFiles(file: ReturnedNode) {
       return {
         name: file.loaderName,
         children: file.files.map(f => ({
           name: f
         }))
       };
-    });
+    }
 
-    const loadedFiles: TreeNode[] = returnedProject.loadedFiles.map(file => {
-      return {
-        name: file.loaderName,
-        children: file.files.map(f => ({
-          name: f
-        }))
-      };
-    });
+    const savedFiles: TreeNode[] = returnedProject.savedFiles.map(file => mapFiles(file));
+    const loadedFiles: TreeNode[] = returnedProject.loadedFiles.map(file => mapFiles(file));
 
     return ({
       projectId: returnedProject.id,
@@ -75,5 +73,14 @@ export class ProjectService {
       savedFiles: savedFiles,
       loadedFiles: loadedFiles
     });
+  }
+
+  private static convertReturnedContextSlicesToContext(contextSlices: ReturnedContextSlice[]) {
+    return contextSlices.map(slice => ({
+      name: slice.name,
+      children: slice.models.map(model => ({
+        name: model
+      }))
+    }));
   }
 }
