@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 
 namespace DxWorks.ScriptBee.Plugin.Api;
 
-public interface IModelLoader
+public interface IModelLoader : IPlugin
 {
     public Task<Dictionary<string, Dictionary<string, ScriptBeeModel>>> LoadModel(List<Stream> fileStreams,
         Dictionary<string, object>? configuration = default, CancellationToken cancellationToken = default);
