@@ -6,12 +6,12 @@ namespace ScriptBee.Models;
 
 public class ProjectModel : IDocument
 {
-    [BsonId] public string Id { get; set; }
-
-    public string Name { get; set; }
+    public string Name { get; set; } = "";
     public DateTime CreationDate { get; set; }
-    public Dictionary<string, List<string>> SavedFiles { get; set; } = new();
-    public Dictionary<string, List<string>> LoadedFiles { get; set; } = new();
-    public string Linker { get; set; }
-    public int LastRunIndex { get; set; }
+    public Dictionary<string, List<FileData>> SavedFiles { get; set; } = new();
+
+    public Dictionary<string, List<FileData>> LoadedFiles { get; set; } = new();
+    public string Linker { get; set; } = "";
+    public Run? LastRun { get; set; }
+    [BsonId] public string Id { get; set; }
 }
