@@ -6,17 +6,8 @@ namespace ScriptBee.Models;
 
 public class RunModel : IDocument
 {
+    public List<Run> Runs { get; set; } = new();
+
     [BsonId]
-    [BsonRepresentation(BsonType.ObjectId)]
-    public string Id { get; set; }
-
-    public int RunIndex { get; set; }
-
-    public string ProjectId { get; set; }
-    public Dictionary<string, List<string>> LoadedFiles { get; set; } = new();
-    public string Linker { get; set; }
-    public string ScriptName { get; set; }
-    public string ConsoleOutputName { get; set; }
-    public List<string> OutputFileNames { get; set; } = new();
-    public string? Errors { get; set; }
+    public string Id { get; set; } = null!;
 }
