@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
 
 namespace ScriptBee.ProjectContext;
-public record FileTreeNode(string name, List<FileTreeNode> children, string filePath, string srcPath);
+
+public record FileTreeNode(string name, List<FileTreeNode>? children, string filePath, string srcPath);
