@@ -9,11 +9,9 @@ public interface IProjectFileStructureManager
 
     public void CreateFile(string projectId, string relativePath, string fileContent);
 
-    public void CreateSrcFile(string projectId, string relativePath, string fileContent);
+    public FileTreeNode CreateSrcFile(string projectId, string relativePath, string fileContent);
 
     public bool FileExists(string projectId, string relativePath);
-
-    public string GetFileContent(string projectId, string relativePath);
 
     public Task<string?> GetFileContentAsync(string projectId, string relativePath);
 
@@ -26,4 +24,5 @@ public interface IProjectFileStructureManager
     public void DeleteFolder(string projectId, string pathToFolder);
 
     public string GetProjectAbsolutePath(string projectId);
+    void SetupFileWatcher(string projectId, string filePath);
 }
