@@ -1,17 +1,28 @@
-import {Component, Input, OnInit} from '@angular/core';
+import { Component, Input } from '@angular/core';
+import { OutputFilesService } from "../../../services/output/output-files.service";
 
 @Component({
   selector: 'app-console-output',
   templateUrl: './console-output.component.html',
   styleUrls: ['./console-output.component.scss']
 })
-export class ConsoleOutputComponent implements OnInit {
+export class ConsoleOutputComponent {
 
-  @Input()
-  consoleOutput: string;
+  // todo add the possibility for realtime updates
 
-  constructor() { }
+  @Input() set outputId(value: string | undefined) {
+    if (!value) {
+      this.consoleOutput = '';
+      return;
+    }
 
-  ngOnInit(): void {
+    this.outputFilesService.fetchOutput(value).subscribe(output => {
+      this.consoleOutput = output;
+    });
+  }
+
+  consoleOutput: string = "";
+
+  constructor(private outputFilesService: OutputFilesService) {
   }
 }
