@@ -1,6 +1,3 @@
-import {ScriptTypes} from './script-types';
-
 export interface CreateScriptDialogData {
-  scriptPath: string;
-  scriptType: ScriptTypes;
+  projectId: string;
 }
