@@ -1,9 +1,10 @@
-import {Component, Inject} from '@angular/core';
-import {MAT_DIALOG_DATA, MatDialogRef} from '@angular/material/dialog';
-import {CreateScriptDialogData} from './create-script-dialog-data';
-import {FileSystemService} from '../../../services/file-system/file-system.service';
-import {ScriptTypes} from './script-types';
-import {ProjectDetailsService} from '../../project-details.service';
+import { Component, Inject } from '@angular/core';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import { CreateScriptDialogData } from './create-script-dialog-data';
+import { ScriptTypes } from './script-types';
+import { Store } from "@ngrx/store";
+import { createScript } from "../../../state/script-tree/script-tree.actions";
+import { selectScriptCreationLoading } from "../../../state/script-tree/script-tree.selectors";
 
 @Component({
   selector: 'app-create-script-dialog',
@@ -15,11 +16,25 @@ export class CreateScriptDialogComponent {
   scriptExists = false;
   types = [ScriptTypes.csharp, ScriptTypes.javascript, ScriptTypes.python];
 
-  constructor(public dialogRef: MatDialogRef<CreateScriptDialogComponent>, private fileSystemService: FileSystemService,
-              @Inject(MAT_DIALOG_DATA) public data: CreateScriptDialogData, private projectDetailsService: ProjectDetailsService) {
-    if (!data || !data.scriptPath || !data.scriptType) {
-      this.data = {scriptPath: '', scriptType: ScriptTypes.csharp};
-    }
+  scriptPath: string = '';
+  scriptType: ScriptTypes = ScriptTypes.csharp;
+  loading: boolean = false;
+
+  constructor(public dialogRef: MatDialogRef<CreateScriptDialogComponent>, @Inject(MAT_DIALOG_DATA) public data: CreateScriptDialogData,
+              private store: Store) {
+    this.store.select(selectScriptCreationLoading).subscribe({
+      next: ({loading, error}) => {
+        if (!loading && this.loading) {
+          if (error) {
+            this.scriptExists = true;
+          } else {
+            this.dialogRef.close();
+          }
+        }
+
+        this.loading = loading;
+      }
+    });
   }
 
   onCancelClick(): void {
@@ -27,16 +42,10 @@ export class CreateScriptDialogComponent {
   }
 
   onOkClick(): void {
-    this.projectDetailsService.project.subscribe(project => {
-      if (project) {
-        this.fileSystemService.createScript(project.projectId, this.data.scriptPath, this.data.scriptType).subscribe(res => {
-          if (res) {
-            this.dialogRef.close(res.filePath);
-          }
-        }, (error: any) => {
-          this.scriptExists = true;
-        });
-      }
-    });
+    this.store.dispatch(createScript({
+      projectId: this.data.projectId,
+      scriptPath: this.scriptPath,
+      scriptType: this.scriptType
+    }));
   }
 }
