@@ -1,6 +1,7 @@
-﻿import {TreeNode} from '../../shared/tree-node';
+﻿import { TreeNode } from '../../shared/tree-node';
 
 export interface FileTreeNode extends TreeNode {
   filePath: string;
   srcPath: string;
+  children?: FileTreeNode[];
 }
