@@ -1,5 +1,4 @@
 export interface OutputFile {
   fileName: string;
-  fileType: string;
-  filePath: string;
+  fileId: string;
 }
