@@ -1,12 +1,12 @@
-import {Injectable} from '@angular/core';
-import {HttpClient, HttpParams} from '@angular/common/http';
-import {contentHeaders} from '../../shared/headers';
+import { Injectable } from '@angular/core';
+import { HttpClient, HttpParams } from '@angular/common/http';
+import { contentHeaders } from '../../shared/headers';
 
 @Injectable({
   providedIn: 'root'
 })
 export class OutputFilesService {
-
+  private getOutputDataUrl = '/api/output';
   private getConsoleOutputApiURL = '/api/output/console';
   private downloadFileApiURL = '/api/output/files/download';
   private downloadAllApiURL = '/api/output/files/downloadAll';
@@ -25,20 +25,29 @@ export class OutputFilesService {
     });
   }
 
-  downloadFile(filePath: string) {
-    return this.http.post(this.downloadFileApiURL, {filePath: filePath}, {
+  downloadFile(id: string, name: string) {
+    return this.http.post(this.downloadFileApiURL, {
+      id,
+      name
+    }, {
       headers: contentHeaders,
       responseType: 'blob'
     });
   }
 
-  downloadAll(projectId: string, runId: string) {
+  downloadAll(projectId: string, runIndex: number) {
     return this.http.post(this.downloadAllApiURL, {
-      projectId: projectId,
-      runId: runId
+      projectId,
+      runIndex
     }, {
       headers: contentHeaders,
       responseType: 'blob'
+    });
+  }
+
+  fetchOutput(outputId: string) {
+    return this.http.get<string>(`${this.getOutputDataUrl}/${outputId}`, {
+      headers: contentHeaders
     });
   }
 }
