@@ -1,10 +1,14 @@
-import {OutputFile} from '../output/output-file';
-
 export interface RunScriptResult {
-  runId: string;
-  projectId: string;
-  consoleOutputName: string;
-  outputFiles: OutputFile[];
-  errors?: string;
+  index: number;
+  linker: string;
+  loadedFiles: Map<string, string[]>;
+  scriptName: string;
+  results: OutputResult[];
+}
+
+export interface OutputResult {
+  id: string;
+  name: string;
+  type: string;
 }
 
