@@ -1,7 +1,7 @@
-import {Injectable} from '@angular/core';
-import {HttpClient} from '@angular/common/http';
-import {contentHeaders} from '../../shared/headers';
-import {RunScriptResult} from './run-script-result';
+import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+import { contentHeaders } from '../../shared/headers';
+import { RunScriptResult } from './run-script-result';
 
 @Injectable({
   providedIn: 'root'
@@ -13,10 +13,11 @@ export class RunScriptService {
   constructor(private http: HttpClient) {
   }
 
-  runScriptFromPath(projectId: string, filePath: string) {
+  runScriptFromPath(projectId: string, filePath: string, language: string) {
     return this.http.post<RunScriptResult>(this.runScriptAPIUrl, {
       projectId: projectId,
-      filePath: filePath
+      filePath: filePath,
+      language: language
     }, {headers: contentHeaders});
   }
 }
