@@ -1,5 +1,6 @@
-import {Injectable} from '@angular/core';
-import {HttpClient} from '@angular/common/http';
+import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+import { UploadModelsResult } from "./upload-models-result";
 
 @Injectable({
   providedIn: 'root'
@@ -11,12 +12,12 @@ export class UploadService {
   constructor(private http: HttpClient) {
   }
 
-  uploadModels(loaderName: string, projectId: string, files: any) {
+  uploadModels(loaderName: string, projectId: string, files: File[]) {
     const formData = new FormData();
     formData.append('loaderName', loaderName);
     files.forEach(file => formData.append('files', file));
     formData.append('projectId', projectId);
-    
-    return this.http.post(this.uploadFilesUrl, formData);
+
+    return this.http.post<UploadModelsResult>(this.uploadFilesUrl, formData);
   }
 }
