@@ -1,12 +1,12 @@
 ﻿export class Upload {
 
   file: File;
-  name: string;
-  url: string;
-  progress: number;
+  name?: string;
+  url?: string;
+  progress?: number;
   createdAt: Date = new Date();
 
-  constructor (file: File) {
+  constructor(file: File) {
     this.file = file;
   }
 }
