@@ -1,3 +1,3 @@
 ﻿namespace ScriptBeeWebApp.Controllers.Arguments;
 
-public record DownloadAll(string ProjectId, string RunId);
+public record DownloadAll(string ProjectId, int RunIndex);
