@@ -6,30 +6,30 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using ScriptBee.Models;
 using ScriptBee.ProjectContext;
+using ScriptBee.Services;
 using ScriptBeeWebApp.Controllers.Arguments;
+using ScriptBeeWebApp.Controllers.DTO;
 using ScriptBeeWebApp.Services;
 
 namespace ScriptBeeWebApp.Controllers;
 
 [ApiControllerRoute]
 [ApiController]
+// todo add tests
 public class ProjectsController : ControllerBase
 {
-    private readonly IProjectManager _projectManager;
-    private readonly IProjectFileStructureManager _projectFileStructureManager;
-    private readonly IProjectModelService _projectModelService;
-    private readonly IFileNameGenerator _fileNameGenerator;
     private readonly IFileModelService _fileModelService;
+    private readonly IProjectFileStructureManager _projectFileStructureManager;
+    private readonly IProjectManager _projectManager;
+    private readonly IProjectModelService _projectModelService;
     private readonly IRunModelService _runModelService;
 
     public ProjectsController(IProjectManager projectManager, IProjectFileStructureManager projectFileStructureManager,
-        IProjectModelService projectModelService, IFileNameGenerator fileNameGenerator,
-        IFileModelService fileModelService, IRunModelService runModelService)
+        IProjectModelService projectModelService, IFileModelService fileModelService, IRunModelService runModelService)
     {
         _projectManager = projectManager;
         _projectFileStructureManager = projectFileStructureManager;
         _projectModelService = projectModelService;
-        _fileNameGenerator = fileNameGenerator;
         _fileModelService = fileModelService;
         _runModelService = runModelService;
     }
@@ -70,19 +70,16 @@ public class ProjectsController : ControllerBase
         }
 
         var contextResult = project.Context.Models.Keys.GroupBy(tuple => tuple.Item1)
-            .Select(grouping => new
-            {
-                name = grouping.Key,
-                children = grouping.Select(t => new
-                {
-                    name = t.Item2
-                })
-            });
+            .Select(grouping => new ReturnedContextSlice(
+                grouping.Key,
+                grouping.Select(tuple => tuple.Item2).ToList()
+            ));
 
         return Ok(contextResult);
     }
 
     [HttpPost]
+    // todo extract validation to separate class
     public async Task<ActionResult<ProjectModel>> CreateProject(CreateProject projectArg,
         CancellationToken cancellationToken)
     {
@@ -103,7 +100,7 @@ public class ProjectsController : ControllerBase
         {
             Id = project.Id,
             Name = project.Name,
-            CreationDate = DateTime.Now,
+            CreationDate = DateTime.Now
         };
 
         await _projectModelService.CreateDocument(projectModel, cancellationToken);
@@ -113,6 +110,7 @@ public class ProjectsController : ControllerBase
         return projectModel;
     }
 
+    // todo reimplement
     [HttpDelete("{projectId}")]
     public async Task<IActionResult> RemoveProject(string projectId, CancellationToken cancellationToken)
     {
@@ -123,16 +121,16 @@ public class ProjectsController : ControllerBase
 
         var projectModel = await _projectModelService.GetDocument(projectId, cancellationToken);
 
+
         if (projectModel != null)
         {
+            var fileIds = new List<string>();
             foreach (var (_, savedFilesNames) in projectModel.SavedFiles)
             {
-                foreach (var fileName in savedFilesNames)
-                {
-                    await _fileModelService.DeleteFile(fileName, cancellationToken);
-                }
+                fileIds.AddRange(savedFilesNames.Select(data => data.Id.ToString()));
             }
 
+            await _fileModelService.DeleteFilesAsync(fileIds, cancellationToken);
             await _projectModelService.DeleteDocument(projectId, cancellationToken);
         }
 
@@ -141,16 +139,17 @@ public class ProjectsController : ControllerBase
         var allRunsForProject = await _runModelService.GetAllRunsForProject(projectId, cancellationToken);
         foreach (var runModel in allRunsForProject)
         {
-            await _fileModelService.DeleteFile(runModel.ScriptName, cancellationToken);
+            await _fileModelService.DeleteFileAsync(runModel.ScriptPath, cancellationToken);
 
-            await _fileModelService.DeleteFile(runModel.ConsoleOutputName, cancellationToken);
+            // todo
+            // await _fileModelService.DeleteFileAsync(runModel.ConsoleOutputName, cancellationToken);
+            //
+            // foreach (var outputFileName in runModel.OutputFileNames)
+            // {
+            //     await _fileModelService.DeleteFileAsync(outputFileName, cancellationToken);
+            // }
 
-            foreach (var outputFileName in runModel.OutputFileNames)
-            {
-                await _fileModelService.DeleteFile(outputFileName, cancellationToken);
-            }
-
-            await _runModelService.DeleteDocument(runModel.Id, cancellationToken);
+            // await _runModelService.DeleteDocument(runModel.Id, cancellationToken);
         }
 
         return Ok("Project removed successfully");
@@ -163,7 +162,7 @@ public class ProjectsController : ControllerBase
 
         if (project == null)
         {
-            return BadRequest($"Project with the given id does not exist");
+            return BadRequest("Project with the given id does not exist");
         }
 
         await _projectModelService.UpdateDocument(projectModel, cancellationToken);
@@ -197,12 +196,9 @@ public class ProjectsController : ControllerBase
         };
     }
 
-    private List<string> ConvertFileNames(IEnumerable<string> files)
+    private List<string> ConvertFileNames(List<FileData> files)
     {
-        return files.Select(file =>
-        {
-            var (_, _, fileName) = _fileNameGenerator.ExtractModelNameComponents(file);
-            return fileName;
-        }).ToList();
+        return files.Select(f => f.Name)
+            .ToList();
     }
 }
