@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using MongoDB.Driver;
@@ -6,6 +7,8 @@ using ScriptBee.Models;
 
 namespace ScriptBeeWebApp.Services;
 
+// todo add tests with test containers
+// todo update docs to reflect changes
 public class RunModelService : MongoService<RunModel>, IRunModelService
 {
     private const string RunsCollectionName = "Runs";
@@ -15,10 +18,39 @@ public class RunModelService : MongoService<RunModel>, IRunModelService
     {
     }
 
-    public async Task<List<RunModel>> GetAllRunsForProject(string projectId, CancellationToken cancellationToken)
+    public async Task<IEnumerable<Run>> GetAllRunsForProject(string projectId,
+        CancellationToken cancellationToken = default)
     {
-        var results = await mongoCollection.FindAsync(run => run.ProjectId == projectId,
-            cancellationToken: cancellationToken);
-        return results.ToList(cancellationToken);
+        var asyncCursor =
+            await mongoCollection.FindAsync(run => run.Id == projectId, cancellationToken: cancellationToken);
+
+        var runModel = await asyncCursor.FirstOrDefaultAsync(cancellationToken);
+
+        return runModel?.Runs ?? Enumerable.Empty<Run>();
+    }
+
+    public async Task AddRun(string projectId, Run run, CancellationToken cancellationToken = default)
+    {
+        var filter = Builders<RunModel>.Filter.Eq(x => x.Id, projectId);
+
+        var asyncCursor = await mongoCollection.FindAsync(filter, cancellationToken: cancellationToken);
+
+        if (await asyncCursor.AnyAsync(cancellationToken: cancellationToken))
+        {
+            var update = Builders<RunModel>.Update.Push(x => x.Runs, run);
+
+            await mongoCollection.UpdateOneAsync(filter, update, cancellationToken: cancellationToken);
+        }
+        else
+        {
+            await mongoCollection.InsertOneAsync(new RunModel
+            {
+                Id = projectId,
+                Runs = new List<Run>
+                {
+                    run
+                }
+            }, cancellationToken: cancellationToken);
+        }
     }
 }
